chore: dev ENV auto login
